requirements.txt
tensorflow
streamlit
opencv-python-headless
numpy
matplotlib
scikit-learn
Pillow